updated pets table - added drop scripts for pets and reminders

+++ db/scripts/createPetsTable.js
@@ -2,7 +2,7 @@ import pg from "pg";
 
 import { query } from "../index.js";
 
-const sqlString = `CREATE TABLE IF NOT EXISTS pets (user_id VARCHAR(10), pet_id VARCHAR(10), name TEXT, species BOOLEAN, breed TEXT, age INT, weight INT);`;
+const sqlString = `CREATE TABLE IF NOT EXISTS pets (user_id VARCHAR(10), pet_id VARCHAR(10), name TEXT, species BOOLEAN, breed TEXT, age INT, weight FLOAT);`;
 
 async function createPetsTable() {
 	const res = await query(sqlString);
